vehicles now rendered in correct locations - need to move

+++ FroggerStarter/Model/GameObject.cs
@@ -55,7 +55,7 @@ namespace FroggerStarter.Model
         /// <value>
         ///     The speed x.
         /// </value>
-        public int SpeedX { get; private set; }
+        public double SpeedX { get; private set; }
 
         /// <summary>
         ///     Gets the y speed of the game object.
@@ -63,7 +63,7 @@ namespace FroggerStarter.Model
         /// <value>
         ///     The speed y.
         /// </value>
-        public int SpeedY { get; private set; }
+        public double SpeedY { get; private set; }
 
         /// <summary>
         ///     Gets the width of the game object.
@@ -133,12 +133,12 @@ namespace FroggerStarter.Model
             this.moveY(this.SpeedY);
         }
 
-        private void moveX(int x)
+        private void moveX(double x)
         {
             this.X += x;
         }
 
-        private void moveY(int y)
+        private void moveY(double y)
         {
             this.Y += y;
         }
@@ -155,7 +155,7 @@ namespace FroggerStarter.Model
         /// </summary>
         /// <param name="speedX">The speed x.</param>
         /// <param name="speedY">The speed y.</param>
-        protected void SetSpeed(int speedX, int speedY)
+        protected void SetSpeed(double speedX, double speedY)
         {
             if (speedX < 0)
             {
